+++ flowkit_ui_backend/models/heartbeat.py
@@ -11,6 +11,8 @@ from typing import Any, Dict, List, Optional  # noqa: F401
 
 from pydantic import AnyUrl, BaseModel, EmailStr, validator, BaseConfig  # noqa: F401
 
+from flowkit_ui_backend.util.config import Settings
+
 # required to make arbitrary datatypes work, see https://github.com/tiangolo/fastapi/issues/2382
 BaseConfig.arbitrary_types_allowed = True
 
@@ -21,25 +23,16 @@ BaseConfig.arbitrary_types_allowed = True
 
 class Heartbeat(BaseModel):
     """
-
     Heartbeat - a model defined in OpenAPI
 
         datetime: The datetime of this Heartbeat [Optional].
-        docker_image: The docker_image of this Heartbeat [Optional].
-        git_branch: The git_branch of this Heartbeat [Optional].
-        git_commit: The git_commit of this Heartbeat [Optional].
-        git_tag: The git_tag of this Heartbeat [Optional].
-        python_package: The python_package of this Heartbeat [Optional].
         python_version: The python_version of this Heartbeat [Optional].
+        config: The runtime configuration object of this app
     """
 
     datetime: Optional[datetime] = None  # this breaks pydantic at least as far as 2.5.2
-    docker_image: Optional[str] = None
-    git_branch: Optional[str] = None
-    git_commit: Optional[str] = None
-    git_tag: Optional[str] = None
-    python_package: Optional[str] = None
     python_version: Optional[str] = None
+    config: Settings
 
     class Config:
         extra = "allow"