+++ tests/test_db.py
@@ -59,7 +59,7 @@ async def test_provision_db(mocker, fresh_pool):
 
 @pytest.mark.asyncio
 async def test_provision_db_force_setup(mocker, fresh_pool):
-    await db.provision_db(force=True, pool=fresh_pool)
+    await db.provision_db(pool=fresh_pool, force=True)
     # TODO: add meaningful tests
 
 
@@ -164,7 +164,7 @@ async def test_load_prepared_sql():
 
 @pytest.mark.asyncio
 async def test_run(fresh_pool):
-    await db.run("SHOW VARIABLES", pool=fresh_pool)
+    await db.run(sql="SHOW VARIABLES", pool=fresh_pool)
     # TODO: add meaningful tests