+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `get_latest_date` unauthenticated endpoint for returning the most recent date of ingested data
 
 ### Changed
+- Configuration is now handled by [pydantic's `BaseSettings`](https://docs.pydantic.dev/1.10/usage/settings/)
 
 ### Fixed
 - Fixed profile page erroring due to a change in the auth0 api